replace prepare.js with prep.js, update docs
