+++ prepare.js
@@ -1,182 +1,292 @@
+const fs = require('fs-extra');
+let argv = require('minimist')(process.argv.slice(2));
+const ipfsAPI = require('ipfs-api');
+const { spawn, execFile } = require('child_process');
+const path = require('path');
 
-const fs = require("fs")
-const argv = require('minimist')(process.argv.slice(2))
-const ipfsAPI = require('ipfs-api')
-const { spawn, execFile } = require('child_process')
-const path = require('path')
+var dir = path.dirname(fs.realpathSync(__filename)) + '/';
+var tmp_dir = path.resolve(process.cwd(), argv.out);
 
-var dir = path.dirname(fs.realpathSync(__filename)) + "/"
+fs.mkdirpSync(tmp_dir);
+// fix pathing so we don't need to worry about what dir we are in.
+const fixPaths = (targetDir, relativePathsArray) => {
+  return relativePathsArray.map(filePath => {
+    let start = path.resolve(process.cwd(), filePath);
+    let localPath = filePath.replace('/workspace/src/', '/');
+    let end = path.resolve(targetDir, localPath);
+    fs.copySync(start, end);
+    return localPath;
+  });
+};
 
-var host = argv["ipfs-host"] || "programming-progress.com"
+const cleanUpAfterInstrumenting = () => {
+  let absPathToDeps = argv.file
+    .map(fileName => {
+      return path.resolve(__dirname, fileName);
+    })
+    .concat(
+      argv._.map(fileName => {
+        let newPath = path.resolve(__dirname, fileName);
+        return newPath;
+      })
+    )
+    .concat(
+      argv._.map(fileName => {
+        let newPath = path.resolve(__dirname, fileName);
+        return newPath.replace(/.js$/, '.wasm');
+      })
+    );
 
-var ipfs = ipfsAPI(host, '5001', {protocol: 'http'})
+  absPathToDeps.forEach(filePath => {
+    fs.unlinkSync(filePath);
+  });
+};
 
-var tmp_dir = "/tmp/emscripten-module-wrapper" + Math.floor(Math.random() * Math.pow(2,32)).toString(32)
+const localizeArgv = argv => {
+  argv._.push(argv._[0].replace(/.js$/, '.wasm'));
+  fixPaths(tmp_dir, argv._);
+  argv._ = [fixPaths(__dirname, argv._)[0]];
 
-var config = []
+  // move files
+  fixPaths(tmp_dir, argv.file);
+  argv.file = fixPaths(__dirname, argv.file);
+  return argv;
+};
+
+argv = localizeArgv(argv);
+
+var config = [];
 
 function readConfig() {
-    try {
-        config = JSON.parse(fs.readFileSync(dir + "../webasm-solidity/node/config.json"))
-    }
-    catch (e) {
-    }
+  try {
+    config = JSON.parse(
+      fs.readFileSync(dir + '../webasm-solidity/node/config.json')
+    );
+  } catch (e) {}
 }
 
-readConfig()
+readConfig();
 
-fs.mkdirSync(tmp_dir)
+var wasm = dir + '../ocaml-offchain/interpreter/wasm';
 
-var wasm = dir + "../ocaml-offchain/interpreter/wasm"
+var prerun = fs.readFileSync(dir + 'pre-run.js');
+var preamble = fs.readFileSync(dir + 'preamble.js');
 
-var prerun = fs.readFileSync(dir+"pre-run.js")
-var preamble = fs.readFileSync(dir+"preamble.js")
-
-function uploadIPFS(fname) {
-    return new Promise(function (cont,err) {
-        fs.readFile(tmp_dir + "/" + fname, function (err, buf) {
-            ipfs.files.add([{content:buf, path:fname}], function (err, res) {
-                cont(res[0])
-            })
-        })
-    })
+function exec(cmd, args) {
+  return new Promise((resolve, reject) => {
+    execFile(cmd, args, { cwd: tmp_dir }, (error, stdout, stderr) => {
+      if (error) {
+        // console.error('error ', error);
+        reject(error);
+      }
+      if (stderr) {
+        // console.error('error ', stderr, args);
+        reject(stderr);
+      }
+      if (stdout) {
+        // console.log('output ', stdout, args);
+      }
+      resolve(stdout);
+    });
+  });
 }
 
-function exec(cmd, args, dr) {
-    return new Promise(function (cont,err) {
-        console.log("exec: ", cmd, args, dr)
-        execFile(cmd, args, {cwd:dr || tmp_dir}, function (error, stdout, stderr) {
-            if (stderr) console.error('error ', stderr, args)
-            if (stdout) console.log('output ', stdout, args)
-/*            if (error) err(error)
-            else cont(stdout) */
-            cont(stdout)
-        })
-    })
-}
+function spawnPromise(cmd, args) {
+  return new Promise((resolve, reject) => {
+    var res = '';
+    const p = spawn(cmd, args, { cwd: tmp_dir });
 
-function spawnPromise(cmd, args, dr) {
-    return new Promise(function (cont,err) {
-        console.log("exec: ", cmd + " " + args.join(" "), dr)
-        var res = ""
-        const p = spawn(cmd, args, {cwd:dr || tmp_dir})
-        
-        p.on('error', (err) => {
-            console.log('Failed to start subprocess.');
-            err(error)
-        });
+    p.on('error', err => {
+      // console.log('Failed to start subprocess.');
+      reject(err);
+    });
 
-        p.stdout.on('data', (data) => {
-            res += data
-            console.log(`stdout: ${data}`);
-        });
+    p.stdout.on('data', data => {
+      res += data;
+      // console.log(`stdout: ${data}`);
+    });
 
-        p.stderr.on('data', (data) => {
-            console.log(`stderr: ${data}`);
-        });
+    p.stderr.on('data', data => {
+      // console.log(`stderr: ${data}`);
+    });
 
-        p.on('close', (code) => {
-            console.log(`child process exited with code ${code}`);
-            cont(res)
-        });
-
-    })
+    p.on('close', code => {
+      // console.log(`child process exited with code ${code}`);
+      resolve(res);
+    });
+  });
 }
 
 function flatten(lst) {
-    return [].concat.apply([],lst)
+  return [].concat.apply([], lst);
 }
 
 function clean(obj, field) {
-    var x = obj[field]
-    if (typeof x == "object") return
-    if (typeof x == "undefined") obj[field] = []
-    else obj[field] = [x]
+  var x = obj[field];
+  if (typeof x == 'object') return;
+  if (typeof x == 'undefined') obj[field] = [];
+  else obj[field] = [x];
 }
 
 async function processTask(fname) {
-    var str = fs.readFileSync(fname, "utf8")
-    str = str.replace(/{{PRE_RUN_ADDITIONS}}/, prerun)
-    
-    if (argv.asmjs) preamble += "\nvar save_stack_top = false;"
-    else preamble += "\nvar save_stack_top = true;"
-    
-    // preamble += "\nvar save_stack_top = true;"
+  var str = fs.readFileSync(path.resolve(tmp_dir, fname), 'utf8');
+  str = str.replace(/{{PRE_RUN_ADDITIONS}}/, prerun);
 
-    str = str.replace(/{{PREAMBLE_ADDITIONS}}/, preamble)
-    str = str.replace(/var exports = null;/, "var exports = null; global_info = info;")
-    str = str.replace(/buffer\.subarray\(/g, "orig_HEAP8.subarray(")
-    str = str.replace(/updateGlobalBufferViews\(\);/, "updateGlobalBufferViews(); addHeapHooks();")
-    str = str.replace(/FS.createStandardStreams\(\);/, "FS.createStandardStreams(); FS.mkdir('/working'); FS.mount(NODEFS, { root: '.' }, '/working'); FS.chdir('/working');")
-    str = str.replace(/Module\[\"noExitRuntime\"\] = true/, 'Module["noExitRuntime"] = false')
-    fs.writeFileSync(tmp_dir + "/prepared.js", 'var source_dir = "' + tmp_dir + '"\n' + str)
+  if (argv.asmjs) preamble += '\nvar save_stack_top = false;';
+  else preamble += '\nvar save_stack_top = true;';
 
-    var wasm_file = fname.replace(/.js$/, ".wasm")
+  // preamble += "\nvar save_stack_top = true;"
 
-    await exec("cp", [wasm_file, tmp_dir + "/" + wasm_file], process.cwd())
+  str = str.replace(/{{PREAMBLE_ADDITIONS}}/, preamble);
+  str = str.replace(
+    /var exports = null;/,
+    'var exports = null; global_info = info;'
+  );
+  str = str.replace(/buffer\.subarray\(/g, 'orig_HEAP8.subarray(');
+  str = str.replace(
+    /updateGlobalBufferViews\(\);/,
+    'updateGlobalBufferViews(); addHeapHooks();'
+  );
+  str = str.replace(
+    /FS.createStandardStreams\(\);/,
+    "FS.createStandardStreams(); FS.mkdir('/working'); FS.mount(NODEFS, { root: '.' }, '/working'); FS.chdir('/working');"
+  );
+  str = str.replace(
+    /Module\[\"noExitRuntime\"\] = true/,
+    'Module["noExitRuntime"] = false'
+  );
+  fs.writeFileSync(
+    tmp_dir + '/prepared.js',
+    'var source_dir = __dirname;\n' + str
+  );
 
-    console.log(argv)
+  var wasm_file = fname.replace(/.js$/, '.wasm');
 
-    clean(argv, "arg")
-    clean(argv, "file")
+  clean(argv, 'arg');
+  clean(argv, 'file');
 
-    console.log(argv)
-    if (argv.analyze) {
-        for (var i = 0; i < argv.file.length; i++) {
-            await exec("cp", [argv.file[i], tmp_dir + "/" + argv.file[i]], process.cwd())
-        }
-        await exec("node", ["prepared.js"].concat(argv.arg))
-        // return
-    }
+  if (argv.analyze) {
+    await exec('node', ['prepared.js'].concat(argv.arg));
+  }
 
-    for (var i = 0; i < argv.file.length; i++) {
-        await exec("cp", [argv.file[i], tmp_dir + "/" + argv.file[i]], process.cwd())
-    }
+  if (argv.asmjs)
+    await exec(wasm, ['-merge', wasm_file, dir + 'filesystem.wasm']);
+  else {
+    await exec(wasm, ['-underscore', wasm_file]);
+    await exec(wasm, [
+      '-merge',
+      'underscore.wasm',
+      dir + 'filesystem-wasm.wasm'
+    ]);
+  }
+  if (argv.analyze)
+    await exec(wasm, ['-add-globals', 'globals.json', 'merge.wasm']);
+  else if (argv.asmjs)
+    await exec(wasm, [
+      '-add-globals',
+      dir + 'globals-asmjs.json',
+      'merge.wasm'
+    ]);
+  else await exec(wasm, ['-add-globals', dir + 'globals.json', 'merge.wasm']);
 
-    if (argv.asmjs) await exec(wasm, ["-merge", wasm_file, dir + "filesystem.wasm"])
-    else {
-        await exec(wasm, ["-underscore", wasm_file])
-        await exec(wasm, ["-merge", "underscore.wasm", dir + "filesystem-wasm.wasm"])
-    }
-    if (argv.analyze) await exec(wasm, ["-add-globals", "globals.json", "merge.wasm"])
-    else if (argv.asmjs) await exec(wasm, ["-add-globals", dir+"globals-asmjs.json", "merge.wasm"])
-    else await exec(wasm, ["-add-globals", dir+"globals.json", "merge.wasm"])
+  var args = flatten(argv.arg.map(a => ['-arg', a]));
+  args = args.concat(flatten(argv.file.map(a => ['-file', a])));
+  if (config.interpreter_args) args = args.concat(config.interpreter_args);
+  if (argv.asmjs) args.push('-asmjs');
+  var result_wasm = 'globals.wasm';
+  var float_memory = 10 * 1024;
 
-    var args = flatten(argv.arg.map(a => ["-arg", a]))
-    args = args.concat(flatten(argv.file.map(a => ["-file", a])))
-    if (config.interpreter_args) args = args.concat(config.interpreter_args)
-    if (argv.asmjs) args.push("-asmjs")
-    var result_wasm = "globals.wasm"
-    var float_memory = 10*1024
+  if (argv.float) {
+    await exec(wasm, ['-shift-mem', float_memory, 'globals.wasm']);
+    await exec(wasm, [
+      '-memory-offset',
+      float_memory,
+      '-int-float',
+      dir + 'softfloat.wasm',
+      'shiftmem.wasm'
+    ]);
+    result_wasm = 'intfloat.wasm';
+    args.push('-memory-offset');
+    args.push(float_memory);
+  }
 
-    if (argv.float) {
-        await exec(wasm, ["-shift-mem", float_memory, "globals.wasm"])
-        await exec(wasm, ["-memory-offset", float_memory, "-int-float", dir + "softfloat.wasm", "shiftmem.wasm"])
-        result_wasm = "intfloat.wasm"
-        args.push("-memory-offset")
-        args.push(float_memory)
-    }
+  if (argv.metering) {
+    var dta = fs.readFileSync(tmp_dir + '/' + result_wasm);
+    const metering = require('wasm-metering');
+    const meteredWasm = metering.meterWASM(dta, {
+      moduleStr: 'env',
+      fieldStr: 'usegas',
+      meterType: 'i64'
+    });
+    result_wasm = 'metered.wasm';
+    var dta = fs.writeFileSync(tmp_dir + '/' + result_wasm, meteredWasm);
+  }
 
-    if (argv.metering) {
-        var dta = fs.readFileSync(tmp_dir + "/" + result_wasm)
-        const metering = require('wasm-metering')
-        const meteredWasm = metering.meterWASM(dta, {
-            moduleStr: "env",
-            fieldStr: "usegas",
-            meterType: 'i64',
-        })
-        result_wasm = "metered.wasm"
-        var dta = fs.writeFileSync(tmp_dir + "/" + result_wasm, meteredWasm)
-    }
+  var mem_size = argv['memory-size'] || '25';
+  var info = await spawnPromise(
+    wasm,
+    [
+      '-m',
+      '-input',
+      '-table-size',
+      '20',
+      '-stack-size',
+      '20',
+      '-memory-size',
+      mem_size,
+      '-wasm',
+      result_wasm
+    ].concat(args)
+  );
+  if (argv.run)
+    await spawnPromise(
+      wasm,
+      [
+        '-m',
+        '-table-size',
+        '20',
+        '-stack-size',
+        '20',
+        '-memory-size',
+        mem_size,
+        '-wasm',
+        result_wasm
+      ].concat(args)
+    );
 
-    var mem_size = argv["memory-size"] || "25"
-    var info = await spawnPromise(wasm, ["-m", "-input", "-table-size", "20", "-stack-size", "20", "-memory-size", mem_size, "-wasm", result_wasm].concat(args))
-    if (argv.run) await spawnPromise(wasm, ["-m", "-table-size", "20", "-stack-size", "20", "-memory-size", mem_size, "-wasm", result_wasm].concat(args))
-    var hash = await uploadIPFS("globals.wasm")
-    console.log("cd", tmp_dir)
-    console.log("Uploaded to IPFS ", hash)
-    fs.writeFileSync("info.json", JSON.stringify({ipfshash: hash.hash, codehash: JSON.parse(info).vm.code, memsize:mem_size}))
+  if (argv['upload-ipfs']) {
+    var host = argv['ipfs-host'] || 'localhost';
+
+    var ipfs = ipfsAPI(host, '5001', { protocol: 'http' });
+
+    const uploadIPFS = fname => {
+      return new Promise(function(cont, err) {
+        fs.readFile(tmp_dir + '/' + fname, function(err, buf) {
+          ipfs.files.add([{ content: buf, path: fname }], function(err, res) {
+            cont(res[0]);
+          });
+        });
+      });
+    };
+
+    var hash = await uploadIPFS('globals.wasm');
+
+    fs.writeFileSync(
+      path.join(tmp_dir, 'info.json'),
+      JSON.stringify(
+        {
+          ipfshash: hash.hash,
+          codehash: JSON.parse(info).vm.code,
+          info: JSON.parse(info),
+          memsize: mem_size
+        },
+        null,
+        2
+      )
+    );
+  }
+
+  cleanUpAfterInstrumenting();
 }
 
-argv._.forEach(processTask)
-
+argv._.forEach(processTask);
